search indexing process coupling
